+++ examples/boxmodelcollisions/lowlist/src/main_lowlistcolls.cpp
@@ -49,8 +49,8 @@
 #include "runcleo/initialconditions.hpp"
 #include "runcleo/runcleo.hpp"
 #include "runcleo/sdmmethods.hpp"
-#include "superdrops/collisionprobs/coalescence.hpp"
-#include "superdrops/collisionprobs/lowlistprob.hpp"
+#include "superdrops/collisions/coalescence.hpp"
+#include "superdrops/collisions/lowlistprob.hpp"
 #include "superdrops/microphysicalprocess.hpp"
 #include "superdrops/motion.hpp"
 #include "zarr/fsstore.hpp"
