+++ libs/superdrops/CMakeLists.txt
@@ -7,7 +7,7 @@ endif()
 set(LIBNAME "superdrops")
 message("${LIBNAME} LIBRARY_SOURCE_DIR: ${CMAKE_CURRENT_SOURCE_DIR}")
 
-add_subdirectory(collisionprobs)
+add_subdirectory(collisions)
 
 # explicitly set library executables path to /lib in top level of build tree
 set(LIBRARY_OUTPUT_PATH  ${CMAKE_BINARY_DIR}/lib)
@@ -26,7 +26,7 @@ set(SOURCES
 add_library("${LIBNAME}" STATIC ${SOURCES})
 
 # Add directories and link libraries for target library
-set(LINKLIBS collisionprobs)
+set(LINKLIBS collisions)
 target_link_libraries(${LIBNAME} PUBLIC "${LINKLIBS}")
 target_link_libraries(${LIBNAME} PUBLIC Kokkos::kokkos)
 target_include_directories(${LIBNAME} PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}")
