+++ libs/superdrops/collisions/collisionkinetics.cpp
@@ -3,7 +3,7 @@
  *
  * ----- CLEO -----
  * File: collisionkinetics.cpp
- * Project: collisionprobs
+ * Project: collisions
  * Created Date: Wednesday 24th January 2024
  * Author: Clara Bayley (CB)
  * Additional Contributors:
