+++ libs/superdrops/collisions/golovinprob.hpp
@@ -4,7 +4,7 @@
  *
  * ----- CLEO -----
  * File: golovinprob.hpp
- * Project: collisionprobs
+ * Project: collisions
  * Created Date: Thursday 9th November 2023
  * Author: Clara Bayley (CB)
  * Additional Contributors:
@@ -25,8 +25,8 @@
  */
 
 
-#ifndef LIBS_SUPERDROPS_COLLISIONPROBS_GOLOVINPROB_HPP_
-#define LIBS_SUPERDROPS_COLLISIONPROBS_GOLOVINPROB_HPP_
+#ifndef LIBS_SUPERDROPS_COLLISIONS_GOLOVINPROB_HPP_
+#define LIBS_SUPERDROPS_COLLISIONS_GOLOVINPROB_HPP_
 
 #include "../../cleoconstants.hpp"
 #include "../superdrop.hpp"
@@ -53,4 +53,4 @@ struct GolovinProb {
                     const double VOLUME) const;
 };
 
-#endif  // LIBS_SUPERDROPS_COLLISIONPROBS_GOLOVINPROB_HPP_
+#endif  // LIBS_SUPERDROPS_COLLISIONS_GOLOVINPROB_HPP_
