+++ libs/superdrops/collisions/hydrodynamicprob.hpp
@@ -4,7 +4,7 @@
  *
  * ----- CLEO -----
  * File: hydrodynamicprob.hpp
- * Project: collisionprobs
+ * Project: collisions
  * Created Date: Thursday 9th November 2023
  * Author: Clara Bayley (CB)
  * Additional Contributors:
@@ -21,8 +21,8 @@
  */
 
 
-#ifndef LIBS_SUPERDROPS_COLLISIONPROBS_HYDRODYNAMICPROB_HPP_
-#define LIBS_SUPERDROPS_COLLISIONPROBS_HYDRODYNAMICPROB_HPP_
+#ifndef LIBS_SUPERDROPS_COLLISIONS_HYDRODYNAMICPROB_HPP_
+#define LIBS_SUPERDROPS_COLLISIONS_HYDRODYNAMICPROB_HPP_
 
 #include <Kokkos_Core.hpp>
 
@@ -70,4 +70,4 @@ struct HydrodynamicProb {
   }
 };
 
-#endif  // LIBS_SUPERDROPS_COLLISIONPROBS_HYDRODYNAMICPROB_HPP_
+#endif  // LIBS_SUPERDROPS_COLLISIONS_HYDRODYNAMICPROB_HPP_
