+++ src/main.cpp
@@ -55,15 +55,15 @@
 #include "runcleo/initialconditions.hpp"
 #include "runcleo/runcleo.hpp"
 #include "runcleo/sdmmethods.hpp"
-#include "superdrops/collisionprobs/breakup.hpp"
-#include "superdrops/collisionprobs/breakup_nfrags.hpp"
-#include "superdrops/collisionprobs/coalbure.hpp"
-#include "superdrops/collisionprobs/coalescence.hpp"
-#include "superdrops/collisionprobs/coalbure_flag.hpp"
-#include "superdrops/collisionprobs/constprob.hpp"
-#include "superdrops/collisionprobs/golovinprob.hpp"
-#include "superdrops/collisionprobs/longhydroprob.hpp"
-#include "superdrops/collisionprobs/lowlistprob.hpp"
+#include "superdrops/collisions/breakup.hpp"
+#include "superdrops/collisions/breakup_nfrags.hpp"
+#include "superdrops/collisions/coalbure.hpp"
+#include "superdrops/collisions/coalescence.hpp"
+#include "superdrops/collisions/coalbure_flag.hpp"
+#include "superdrops/collisions/constprob.hpp"
+#include "superdrops/collisions/golovinprob.hpp"
+#include "superdrops/collisions/longhydroprob.hpp"
+#include "superdrops/collisions/lowlistprob.hpp"
 #include "superdrops/condensation.hpp"
 #include "superdrops/microphysicalprocess.hpp"
 #include "superdrops/motion.hpp"
